+++ hydra-tui/src/Hydra/TUI/Options.hs
@@ -28,7 +28,7 @@ data Options = Options
   { hydraNodeHost :: Host
   , cardanoConnection :: Either FilePath SocketPath
   , cardanoNetworkId :: NetworkId
-  , cardanoSigningKey :: FilePath
+  , fundsSigningKey :: FilePath
   -- ^ User key used by the tui client to commit
   , fuelVerificationKey :: Maybe FilePath
   -- ^ Optional verification key of the hydra-node's internal wallet, used only
@@ -101,10 +101,10 @@ parseNodeHost =
 parseCardanoSigningKey :: Parser FilePath
 parseCardanoSigningKey =
   strOption
-    ( long "cardano-signing-key"
+    ( long "funds-signing-key"
         <> short 'k'
         <> metavar "FILE"
-        <> help "The path to the user signing key file used for selecting UTxO and signing a commit transaction. This file uses the same 'TextEnvelope' format as cardano-cli."
+        <> help "The path to the funds signing key file used for selecting UTxO and signing a commit transaction. This file uses the same 'TextEnvelope' format as cardano-cli."
         <> value "me.sk"
         <> showDefault
     )